% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R, R/random_forest.R
\name{bagging}
\alias{bagging}
\title{Function that implement the algorithm of Bagging}
\usage{
bagging(fweak, data, model_num, reg)

bagging(fweak, data, model_num, reg)
}
\arguments{
\item{fweak}{- function that generates estimate from weak model based on input}

\item{data}{- list of data that fweak need}

\item{model_num}{- the number of weak models you want to train and combine}

\item{reg}{- logical value, true if the weak model is doing regression, otherwise it's doing classification}

\item{fweak}{- function that generates estimate from weak model based on input}

\item{data}{- list of data that fweak need}

\item{model_num}{- the number of weak models you want to train and combine}

\item{reg}{- logical value, true if the weak model is doing regression, otherwise it's doing classification}
}
\value{
bagging(fweak, data, model_num, reg)

outputs bagging_fit1(fweak, data)
}
\description{
Function that implement the algorithm of Bagging

Function that implement the algorithm of Bagging
}
\examples{
fweak <- function(x, y){
  lm(y ~ x)$coefficients
}
data <- list(x = matrix(rnorm(1000), 200, 5))
data$y <- data$x \%*\% rnorm(5)
model_num <- 100; reg <- T
bagging(fweak, data, model_num, reg)
fweak <- function(x, y){
  lm(y ~ x)$coefficients
}
data <- list(x = matrix(rnorm(1000), 200, 5))
data$y <- data$x \%*\% rnorm(5)
bagging_fit1(fweak, data)
}
