% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting.R
\name{boosting_fit1}
\alias{boosting_fit1}
\title{Function that implement one resample of Boosting in regression}
\usage{
boosting_fit1(fweak, data)
}
\arguments{
\item{fweak}{- function that generates estimate from weak model based on input}

\item{data}{- list of data that fweak need}
}
\value{
outputs boosting_fit1(fweak, data)
}
\description{
Function that implement one resample of Boosting in regression
}
\examples{
fweak <- function(x, y){
  lm(y ~ x)$coefficients
}
data <- list(x = matrix(rnorm(1000), 200, 5))
data$y <- data$x \%*\% rnorm(5)
boosting_fit1(fweak, data)
}
