% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleFuns.R
\name{Comb_parallel}
\alias{Comb_parallel}
\title{Function that combines results from weak models in parallel}
\usage{
Comb_parallel(multi_est, weights, reg)
}
\arguments{
\item{multi_est}{- list containing multiple estimates from weak models}

\item{weights}{- nonegative vector that combines the multiple estimates}

\item{reg}{- logical value, true if the weak model is doing regression, otherwise it's doing classification}
}
\value{
outputs Comb_parallel(multi_est, weights),
}
\description{
Function that combines results from weak models in parallel
}
\examples{
multi_est <- list(1, 2, 3)
weights <- c(0.2, 0.4, 0.4)
reg <- T
Comb_parallel(multi_est, weights, reg)
}
