% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{dt_reg}
\alias{dt_reg}
\title{Function that builds weak model on Decision Tree for regression task}
\usage{
dt_reg(x, y)
}
\arguments{
\item{x}{- input independent variables x for the training}

\item{y}{- input dependent variable y for the traisning}
}
\value{
dt_reg(x ,y)
}
\description{
Function that builds weak model on Decision Tree for regression task
}
\examples{
x <- matrix(rnorm(4000), 200, 20)
beta <- rnorm(5)
y <- x[, 1:length(beta)] \%*\% beta + rnorm(200)
dt_reg(x, y)
}
