% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_define.R
\name{Ensemble_define}
\alias{Ensemble_define}
\title{Function that combines results from weak models in parallel}
\usage{
Ensemble_define(data, parallel, fit_fun, fweak, model_num)
}
\arguments{
\item{data}{- list of data that fweak need, including x and y when parallel is true and x, y and last_est when parallel is False}

\item{parallel}{- logical value, true if its results will be combined in parallel}

\item{fit_fun}{- function that implement one fweak in the user-defined Ensemble Learning: if parallel is False, it should return both the fitted value and the updated last_est}

\item{fweak}{- function that generates estimate from weak model based on input}

\item{model_num}{- the number of weak models you want to train and combine}
}
\value{
outputs Ensemble_define(data, parallel, fit_fun, fweak, model_num),
}
\description{
Function that combines results from weak models in parallel
}
\examples{
data <- list(x = matrix(rnorm(1000), 200, 5))
data$y <- data$x \%*\% rnorm(5) + rnorm(200, 0, 3)
data$last_est <- rep(0, 5)
parallel <- FALSE
fit_fun <- function(data, fweak){
  fit <- graboo_fit1(data = data, fweak = fweak)
  last_est <- fit(diag(rep(1, ncol(data$x))))
  list(model = fit, last_est = last_est)
}
fweak <- graboo_reg
model_num <- 100
Ensemble_define(data, parallel, fit_fun, fweak, model_num)
}
