% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleFuns.R
\name{Ensemblelearn}
\alias{Ensemblelearn}
\title{Function that implement the algorithm of Ensemble learning}
\usage{
Ensemblelearn(fweak, data, model_num, reg, model)
}
\arguments{
\item{fweak}{- function that generates estimate from weak model based on input}

\item{data}{- list of data that fweak need}

\item{model_num}{- the number of weak models you want to train and combine}

\item{reg}{- logical value, true if the weak model is doing regression, otherwise it's doing classification}

\item{model}{- character value that specify which weak model you want to use}
}
\value{
bagging(fweak, data, model_num, reg)
}
\description{
Function that implement the algorithm of Ensemble learning
}
\examples{
fweak <- function(x, y){
  lm(y ~ -1 + x)$coefficients
}
data <- list(x = matrix(rnorm(1000), 200, 5))
data$y <- data$x \%*\% rnorm(5)
model_num <- 100; reg <- T; model <- "bagging"
Ensemblelearn(fweak, data, model_num, reg, model)
}
