% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleFuns.R
\name{fit_model}
\alias{fit_model}
\title{Function that gets results from weak model}
\usage{
fit_model(fweak, parallel, data)
}
\arguments{
\item{fweak}{- function that generates estimate from weak model based on input}

\item{parallel}{- logical value, true if its results will be combined in parallel}

\item{data}{- list of data that fweak need}
}
\value{
The new trained results based on the weak model.
}
\description{
Function that gets results from weak model
}
\examples{
fweak <- function(x, y){
  lm(y ~ -1 + x)$coefficients
}
data <- list(x = matrix(rnorm(1000), 200, 5))
parallel <- TRUE
data$y <- data$x \%*\% rnorm(5) + rnorm(200, 0, 3)
fit_model(fweak, parallel, data)
}
