% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_boosting.R
\name{graboo_fit1}
\alias{graboo_fit1}
\title{Function that implement one weak model step of Gradient Boosting in regression}
\usage{
graboo_fit1(data, fweak = graboo_reg, loss = mse, eta = 0.1)
}
\arguments{
\item{data}{- list of data that fweak need}

\item{fweak}{- function that generates estimate from weak model based on input, its default value is graboo_reg}

\item{loss}{- the loss function used, its default value is the mean of the square error}

\item{eta}{- the step size we use to update the total estimate each time, its default value is 0.1}
}
\value{
A trained model function based on the one implementation of the weak model, whose input is the independent variables.
}
\description{
Function that implement one weak model step of Gradient Boosting in regression
}
\examples{
data <- list(x = matrix(rnorm(1000), 200, 5))
data$y <- data$x \%*\% rnorm(5) + rnorm(200, 0, 3)
data$last_est <- rep(0, ncol(data$x))
graboo_fit1(data)
}
