% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_boosting.R
\name{mse}
\alias{mse}
\title{Function that calculate the mse of dependent variable and its estimate}
\usage{
mse(y, y_hat)
}
\arguments{
\item{y}{- the true dependent variable}

\item{y_hat}{- the estimate of the dependent variable}
}
\value{
outputs mse(y, y_hat)
}
\description{
Function that calculate the mse of dependent variable and its estimate
}
\examples{
y <- rnorm(100); y_hat <- rep(0, 100)
mse(y, y_hat)
}
