% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_boosting.R
\name{nega_gra}
\alias{nega_gra}
\title{Function that calculate the negative gradient of given loss function}
\usage{
nega_gra(loss, x, y, coef, eps = 0.001)
}
\arguments{
\item{loss}{- the loss function used}

\item{x}{- input independent variables x for the training}

\item{y}{- the true dependent variable}

\item{coef}{- the estimate of the coefficients in regression}

\item{eps}{- the small increase used in the calculation of gradient}
}
\value{
The negative gradient value of the given loss function.
}
\description{
Function that calculate the negative gradient of given loss function
}
\examples{
x <- matrix(rnorm(4000), 200, 20)
beta <- rnorm(5)
y <- x[, 1:length(beta)] \%*\% beta + rnorm(200)
coef <- lm(y ~ -1 + x)$coefficients
loss <- mse
nega_gra(loss, x, y, coef)
}
