% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleFuns.R
\name{prediction}
\alias{prediction}
\title{Function that does prediction based on the fitted models for the new coming data}
\usage{
prediction(x, model_train, parallel)
}
\arguments{
\item{x}{- the new coming data that we want to do prediction on}

\item{model_train}{- the list of models trained on the training data}

\item{parallel}{- logical value, true if its results will be combined in parallel}
}
\value{
outputs prediction(x, model_train, parallel)
}
\description{
Function that does prediction based on the fitted models for the new coming data
}
\examples{
x = matrix(rnorm(1000), 200, 5)
y <- x \%*\% rnorm(5) + rnorm(200, 0, 3)
parallel <- TRUE
coef <- lm(y ~ -1 + x)$coefficients
model_train <- list()
model_train[[1]] <- function(x){
  x <- matrix(x, ncol = length(coef))
  x \%*\% coef
}
x_new <- matrix(rnorm(5), 1, 5)
prediction(x_new, model_train, parallel)
}
