% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{Randomforest}
\alias{Randomforest}
\title{Function that implement the algorithm of Ensemble learning in Random Forest}
\usage{
Randomforest(data, model_num, reg, fweak = dt_reg)
}
\arguments{
\item{data}{- list of data that fweak need}

\item{model_num}{- the number of weak models you want to train and combine}

\item{reg}{- logical value, true if the weak model is doing regression, otherwise it's doing classification}

\item{fweak}{- function that generates estimate from weak model based on input}
}
\value{
bagging(fweak, data, model_num, reg)
}
\description{
Function that implement the algorithm of Ensemble learning in Random Forest
}
\examples{
data <- list(x = matrix(rnorm(1000), 200, 5))
data$y <- data$x \%*\% rnorm(5)
model_num <- 100; reg <- T
Randomforest(data, model_num, reg)
}
