% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{rf_cla}
\alias{rf_cla}
\title{Function that builds weak model on Decision Tree for classification task}
\usage{
rf_cla(x, y)
}
\description{
Function that builds weak model on Decision Tree for classification task
}
