% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{rf_reg}
\alias{rf_reg}
\title{Function that builds weak model on Decision Tree for regression task}
\usage{
rf_reg(x, y)
}
\description{
Function that builds weak model on Decision Tree for regression task
}
