% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost.R
\name{Adaboost}
\alias{Adaboost}
\title{Function that implement the algorithm of Ensemble learning in AdaBoost}
\usage{
Adaboost(fweak, data, model_num)
}
\arguments{
\item{fweak}{- function that generates estimate from weak model based on input}

\item{data}{- list of data that fweak need including x, y and last_est}

\item{model_num}{- the number of weak models you want to train and combine}
}
\value{
Adaboost(fweak, data, model_num)
}
\description{
Function that implement the algorithm of Ensemble learning in AdaBoost
}
\examples{
fweak <- function(x, y, last_est){
  lm(y ~ -1 + x, weights = last_est)$coefficients
}
data <- list(x = matrix(rnorm(1000), 200, 5))
data$y <- data$x \%*\% rnorm(5) + rnorm(200, 0, 3)
data$last_est <- rep(1/length(data$y), length(data$y))
model_num <- 100
Adaboost(fweak, data, model_num)
}
