% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{Bagging}
\alias{Bagging}
\title{Function that implement the algorithm of Ensemble learning in Bagging}
\usage{
Bagging(fweak, data, model_num, reg)
}
\arguments{
\item{fweak}{- function that generates estimate from weak model based on input}

\item{data}{- list of data that fweak need}

\item{model_num}{- the number of weak models you want to train and combine}

\item{reg}{- logical value, true if the weak model is doing regression, otherwise it's doing classification}
}
\value{
bagging(fweak, data, model_num, reg)
}
\description{
Function that implement the algorithm of Ensemble learning in Bagging
}
\examples{
fweak <- function(x, y){
  lm(y ~ -1 + x)$coefficients
}
data <- list(x = matrix(rnorm(1000), 200, 5))
data$y <- data$x \%*\% rnorm(5)
model_num <- 100; reg <- T
Bagging(fweak, data, model_num, reg)
}
