% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_boosting.R
\name{Graboo}
\alias{Graboo}
\title{Function that implement the algorithm of Ensemble learning in Gradient Boosting}
\usage{
Graboo(data, model_num, reg, loss = mse, eta = 0.1,
  fweak = graboo_reg)
}
\arguments{
\item{data}{- list of data that fweak need including x, y and last_est}

\item{model_num}{- the number of weak models you want to train and combine}

\item{reg}{- logical value, true if the weak model is doing regression, otherwise it's doing classification}

\item{loss}{- the loss function used, its default value is the mean of the square error}

\item{eta}{- the step size we use to update the total estimate each time, its default value is 0.1}

\item{fweak}{- function that generates estimate from weak model based on input, its default value is graboo_reg}
}
\value{
Graboo(fweak, data, model_num, reg)
}
\description{
Function that implement the algorithm of Ensemble learning in Gradient Boosting
}
\examples{
data <- list(x = matrix(rnorm(1000), 200, 5))
data$y <- data$x \%*\% rnorm(5)
data$last_est <- rep(0, 5)
model_num <- 100; reg <- T
Graboo(data, model_num, reg)
}
