% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{Randomforest}
\alias{Randomforest}
\title{Function that implement the algorithm of Ensemble learning in Random Forest}
\usage{
Randomforest(data, model_num, fweak = dt_reg)
}
\arguments{
\item{data}{- list of data that fweak need}

\item{model_num}{- the number of weak models you want to train and combine}

\item{fweak}{- function that generates estimate from weak model based on input}
}
\value{
Randomforest(data, model_num)
}
\description{
Function that implement the algorithm of Ensemble learning in Random Forest
}
\examples{
x <- matrix(rnorm(4000), 200, 20)
beta <- rnorm(5)
y <- x[, 1:length(beta)] \%*\% beta + rnorm(200)
data <- list(x = x, y = y)
model_num <- 100
Randomforest(data, model_num)
}
