% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{randomforest_fit1}
\alias{randomforest_fit1}
\title{Function that implement one weak model of Random Forest based on the resample of the features}
\usage{
randomforest_fit1(data, fweak = dt_reg, fea_len = NULL)
}
\arguments{
\item{data}{- list of data that fweak need}

\item{fweak}{- function that generates estimate from weak model based on input, its default value is dt_reg}

\item{fea_len}{- the number of features we sample each time, its default value is null}
}
\value{
A trained model function based on the one implementation of the weak model, whose input is the independent variables.
}
\description{
Function that implement one weak model of Random Forest based on the resample of the features
}
\examples{
x <- matrix(rnorm(4000), 200, 20)
beta <- rnorm(5)
y <- x[, 1:length(beta)] \%*\% beta + rnorm(200)
data <- list(x = x, y = y)
randomforest_fit1(data)
}
